tests/wpt/tests/speculation-rules/prerender/blob_object_url.https.html

<!DOCTYPE html>
<title>Same-origin prerendering page can create a url for the given
objects</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const {exec} = await create_prerendered_page(t);

  const result = await exec(async () => {
    const blob_contents = "test blob contents";
    const blob = new Blob([blob_contents]);
    const url = URL.createObjectURL(blob);
    const fetched_content = await fetch(url).then(response => response.text());
    URL.revokeObjectURL(url);
    return fetched_content === blob_contents ? "PASS" : "FAIL";
  });

  // Start prerendering a page that attempts to create a url for a blob.
  assert_equals(
    result, "PASS",
    'prerendering page should be able to create a url for blob and fetch it.');
}, 'prerendering page should be able create url');

</script>
